+++ src/com/facebook/buck/android/BUCK
@@ -31,7 +31,7 @@ java_immutables_library(
     '//third-party/java/guava:guava',
     '//src/com/facebook/buck/cxx:platform',
     '//src/com/facebook/buck/jvm/java:classhash',
-    '//src/com/facebook/buck/jvm/java:nativelinkable',
+    '//src/com/facebook/buck/jvm/core:nativelinkable',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/rules:build_rule',
     '//src/com/facebook/buck/util:exceptions',
@@ -127,8 +127,8 @@ java_immutables_library(
     '//src/com/facebook/buck/file:rules',
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
+    '//src/com/facebook/buck/jvm/core:nativelinkable',
     '//src/com/facebook/buck/jvm/java:classhash',
-    '//src/com/facebook/buck/jvm/java:nativelinkable',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:steps',
     '//src/com/facebook/buck/jvm/java:support',
