+++ src/com/facebook/buck/cxx/BUCK
@@ -55,7 +55,7 @@ java_immutables_library(
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:nativelinkable',
+    '//src/com/facebook/buck/jvm/core:nativelinkable',
     '//src/com/facebook/buck/log:api',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/parser:rule_pattern',
