+++ src/com/facebook/buck/jvm/core/BUCK
@@ -1,0 +1,14 @@
+java_library(
+  name = 'nativelinkable',
+  srcs = [
+    'JavaNativeLinkable.java',
+  ],
+  deps = [
+    '//src/com/facebook/buck/cxx:platform',
+    '//src/com/facebook/buck/rules:build_rule',
+    '//third-party/java/guava:guava',
+  ],
+  visibility = [
+    'PUBLIC',
+  ],
+)
