+++ test/com/facebook/buck/cxx/BUCK
@@ -7,7 +7,7 @@ java_library(
     '//src/com/facebook/buck/cxx:platform',
     '//src/com/facebook/buck/cxx:rules',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:nativelinkable',
+    '//src/com/facebook/buck/jvm/core:nativelinkable',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/python:rules',
     '//src/com/facebook/buck/python:support',
@@ -44,7 +44,7 @@ java_test(
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:nativelinkable',
+    '//src/com/facebook/buck/jvm/core:nativelinkable',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/parser:rule_pattern',
     '//src/com/facebook/buck/python:rules',
